Renamed export::Formatter use clauses to be '__private::Formatter' because it's different now apparently
