+++ src/model/user/profile.rs
@@ -1,5 +1,5 @@
 use crate::model::user::{Color, ModLevel};
-use serde::{export::Formatter, Deserialize, Serialize};
+use serde::{__private::Formatter, Deserialize, Serialize};
 use std::{borrow::Cow, fmt::Display};
 
 #[derive(Debug, Eq, PartialEq, Clone, Serialize, Deserialize)]
