+++ src/response.rs
@@ -14,7 +14,7 @@ use crate::{
     },
     DeError, HasRobtopFormat,
 };
-use serde::export::Formatter;
+use serde::__private::Formatter;
 use std::fmt::Display;
 
 // Since NoneError is not stabilized, we cannot do `impl From<NoneError> for ResponseError<'_>`, so
