+++ src/serde/ser/error.rs
@@ -1,4 +1,4 @@
-use serde::export::Formatter;
+use serde::__private::Formatter;
 use std::{fmt::Display, io};
 
 /// Errors that can occur during serialization
